feat: Type System
Regorus's type system is a first of many features intended to
enable type checking and various other constraints on Rego policies.

The type system is insipired from:
   - JSON schema
   - Bicep

The type system is a subset of JSON schema that are useful
for type-checking policies written for cloud resources.
It is expected to be also useful for most general scenarious as well
since the Rego `input` generally refers to an instance of a type
that can be expressed in many strongly typed programming languages.

Signed-off-by: Anand Krishnamoorthi <anakrish@microsoft.com>